Code Complete - Technically
Ok, technically the code is complete. Doesn't compile and I'm sure, since I've been working on it late at night it has more holes in it than a block of cheese. Check this in, then get it compiling with a small driver app as a single 'c' app. Code review and function descriptions needed.